mbti/需求/星座map搭建，朋友主页UI调整，头像跳转功能实现

+++ XDPNetwork/src/main/java/com/spread/xdpnetwork/network/service/LoginService.kt
@@ -50,7 +50,7 @@ interface LoginService {
     fun searchTagWordByTypeId(@Query("current") current: Int,@Query("typeId") typeId: Int,@Query("keyword") keyword:String): Call<BlogsResponse>
 
     @GET("/wz/blog/queryOnesBlog")
-    fun queryOnesBlog(@Query("current") current: Int,@Query("userId") userId: Int): Call<BlogsResponse>
+    fun queryOnesBlog(@Query("current") current: Int,@Query("userId") userId: Long): Call<BlogsResponse>
 
     @GET("/wz/user/otherUser/{userId}")
     fun queryOther(@Path("userId") userId:Long):Call<UserResponse>
